requirements.txt
einops
imgaug
matplotlib
numpy
opencv_python
Pillow
scipy
scikit-image
torchvision
torch
torchaudio
torchinfo
hdbscan
pandas
tqdm
pomegranate
ptflops
termcolor
pyyaml
duckdb
timm==0.9.16